+++ sdk/daml-lf/data/BUILD.bazel
@@ -64,7 +64,8 @@ da_scala_test(
     scalacopts = lf_scalacopts,
     deps = [
         ":data",
-        "@maven//:org_bouncycastle_bcprov_jdk15on",
+        "//libs-scala/crypto:crypto-tests",
+        "@maven//:com_google_protobuf_protobuf_java",
     ],
 )
